Move Lines into own namespace
