+++ library/Vanilla/Quill/Blots/Lines/SpoilerLineBlot.php
@@ -5,7 +5,7 @@
  * @license https://opensource.org/licenses/GPL-2.0 GPL-2.0
  */
 
-namespace Vanilla\Quill\Blots;
+namespace Vanilla\Quill\Blots\Lines;
 
 class SpoilerLineBlot extends AbstractLineBlot {
